refactor: 포트폴리오 작성 step 6 프로젝트 작성 모달 TestCreateProjectModal -> CreateProject 컴포넌트 반영
