+++ src/components/createPortfolio/createStep/Step06Project.tsx
@@ -11,7 +11,7 @@ import { deleteProject } from '@src/apis/project';
 
 import * as S from '@src/components/common/createPortfolio/createStepStyles';
 import ProjectItem from '@src/components/project/ProjectItem';
-import TestCreateProjectModal from '../TestCreateProjectModal';
+import CreateProject from '@src/components/myProject/CreateProject';
 import TitleTextLabel from '@src/components/common/createPortfolio/TitleTextLabel';
 import NextStepButton from '@src/components/common/createPortfolio/NextStepButton';
 import PrevStepButton from '@src/components/common/createPortfolio/PrevStepButton';
@@ -25,12 +25,10 @@ const Step06Project = ({ onNextButtonClick, onPrevButtonClick }: CreatePortfolio
   const queryClient = useQueryClient();
   const projectData = queryClient.getQueryData('projectData') as ProjectDataType | null;
 
+  const isProjectListExist = projectList.length !== 0;
+
   const handleModal = () => {
     setIsModalOpen(!isModalOpen);
-  };
-
-  const keepModalWindow = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.stopPropagation();
   };
 
   const onClickdeleteProject = (id: number) => {
@@ -42,6 +40,8 @@ const Step06Project = ({ onNextButtonClick, onPrevButtonClick }: CreatePortfolio
   };
 
   useEffect(() => {
+    console.log('프로젝트 데이터 있니? => ', projectData);
+
     if (projectData) {
       setIsProjectExist(true);
       setProjectList(prevProject => [...prevProject, projectData]);
@@ -50,9 +50,13 @@ const Step06Project = ({ onNextButtonClick, onPrevButtonClick }: CreatePortfolio
   }, [projectData]);
 
   useEffect(() => {
-    projectList.length !== 0 && setIsProjectExist(true);
-    projectList.length === 0 && setIsProjectExist(false);
+    console.log('이건 실행되냐??');
+
+    isProjectListExist ? setIsProjectExist(true) : setIsProjectExist(false);
   }, [projectList]);
+
+  console.log('projectData => ', projectData);
+  console.log('projectList => ', projectList);
 
   const title = '프로젝트를 추가하세요';
   const description =
@@ -64,7 +68,7 @@ const Step06Project = ({ onNextButtonClick, onPrevButtonClick }: CreatePortfolio
       <StAddProjectContainer>
         <StAddProjectButton onClick={handleModal}>+ 프로젝트 추가</StAddProjectButton>
         <StProjectsList isprojectexist={`${isProjectExist}`}>
-          {projectList?.length !== 0 &&
+          {isProjectListExist &&
             projectList.map(project => (
               <ProjectItem
                 key={project.id}
@@ -90,7 +94,7 @@ const Step06Project = ({ onNextButtonClick, onPrevButtonClick }: CreatePortfolio
 
       {isModalOpen && (
         <StBackground onClick={handleModal}>
-          <TestCreateProjectModal setIsModalOpen={setIsModalOpen} onClick={keepModalWindow} />
+          <CreateProject showModal1={isModalOpen} setShowModal1={setIsModalOpen} />
         </StBackground>
       )}
     </S.Container>
